made a "impl" folder in tests, created interfaces and used bean of them instead of classes
