+++ src/main/java/pl/wiktor/management/user/model/mapper/Mapper.java
@@ -1,6 +1,6 @@
 package pl.wiktor.management.user.model.mapper;
 
-import pl.wiktor.management.user.model.dto.response.AccountDTO;
+import pl.wiktor.management.user.model.dto.request.AccountDTO;
 import pl.wiktor.management.user.model.entity.Account;
 
 public class Mapper {
@@ -8,7 +8,7 @@ public class Mapper {
         Account account = new Account();
         account.setId(accountDTO.getId());
         account.setLogin(accountDTO.getLogin());
-        account.setPassword(account.getPassword());
+        account.setPassword(accountDTO.getPassword());
         return account;
     }
 }
