+++ src/main/java/pl/wiktor/management/user/service/ActiveUserService.java
@@ -1,65 +1,8 @@
 package pl.wiktor.management.user.service;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import pl.wiktor.management.user.model.dto.response.AccountDTO;
-import pl.wiktor.management.user.model.entity.Account;
-import pl.wiktor.management.user.model.enums.TableSearcher;
-import pl.wiktor.management.user.exception.AccountLoginException;
-import pl.wiktor.management.user.exception.AccountPasswordException;
-import pl.wiktor.management.user.model.mapper.Mapper;
-import pl.wiktor.management.user.repository.AccountRepositoryImpl;
-import pl.wiktor.management.user.repository.ActiveUserRepository;
+import pl.wiktor.management.user.model.dto.request.AccountDTO;
 
-import java.util.UUID;
-
-@Service
-public class ActiveUserService {
-    private ActiveUserRepository activeUserRepository;
-    private AccountRepositoryImpl accountRepository;
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-    @Autowired
-    public ActiveUserService(ActiveUserRepository activeUserRepository, AccountRepositoryImpl accountRepository) {
-        this.activeUserRepository = activeUserRepository;
-        this.accountRepository = accountRepository;
-    }
-
-    public String login(AccountDTO accountDTO) {
-        Account account = Mapper.map(accountDTO);
-        String token = UUID.randomUUID().toString();
-        try {
-            loginIsInDb(account.getLogin());
-            loginAndPasswordCorrect(account);
-            activeUserRepository.login(account.getLogin(), token);
-            logger.info("Account successful logged in");
-            return token;
-        } catch (AccountPasswordException | AccountLoginException e) {
-            logger.error(e.getMessage());
-        }
-        return null;
-    }
-
-    public void logout(String token){
-        if(activeUserRepository.isAccountLoggedIn(TableSearcher.ActiveAccountByToken, token)){
-            activeUserRepository.logout(token);
-            logger.info("Account logged out");
-        } else {
-            logger.error("Account isn't logged in, can't be logged out");
-        }
-    }
-
-    private void loginIsInDb(String login) throws AccountLoginException {
-        if( !accountRepository.isLoginInDb(login)){
-            throw new AccountLoginException("Account isn't registered");
-        }
-    }
-
-    private void loginAndPasswordCorrect(Account account) throws AccountPasswordException{
-        if( !accountRepository.isAccountExistWithPassword(account.getLogin(), account.getPassword())){
-            throw new AccountPasswordException("Password doesn't match");
-        }
-    }
+public interface ActiveUserService {
+    String login(AccountDTO accountDTO);
+    void logout(String token);
 }
